+++ Essential_Feed/Essential_Feed/FeedCache/FeedStore.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+public protocol FeedStore {
+    typealias DeletionCompletion = (Error?) -> Void
+    typealias InsetionCompletion = (Error?) -> Void
+    
+    func deleteCachedFeed(completion: @escaping DeletionCompletion)
+    func insert(_ items: [FeedItem], timestamp: Date, completion: @escaping InsetionCompletion)
+}
